made new settings apply to all button LEDs on lights bigger than 8K ROM
This replaces "USE_CONFIGURABLE_RGB_VOLTAGE_LEVELS"
with "USE_AUX_THRESHOLD_CONFIG", which controls the brightness of
button LEDs while the main LEDs are on,
and during post-off voltage display.

Same basic concept, but works on single-color LEDs too,
and lets the user finally configure POVD thresholds.

The code for this is a bit messy, but the aux LED code as a whole
is pretty messy since it wasn't designed for the things it does now.
The entire thing needs a refactor or rewrite someday.  But not today.
For now, this is just enough to make the pull request
cover more use cases before merging into trunk.

I've tested it on a variety of lights, but am not yet entirely
comfortable with it.  However, it worked on at least these:
- 1-color button LED, no RGB
- front RGB, 1-color button LED
- front RGB, hardwired also to RGB button
- RGB button, no other aux

These may need extra changes,
and may have extra config options which do nothing...
- front RGB, no button LED
- 1-color front aux, no button LED
- no aux at all
- attiny85 lights (some could theoretically support the new options,
  but none even try)